# LegalHawk.Backend.Client/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "LegalHawk.Backend",
    "version": "1.0"
  },
  "paths": {
    "/api/v1/legal-contracts": {
      "get": {
        "tags": [
          "LegalContracts"
        ],
        "operationId": "GetLegalContractsAsync",
        "parameters": [
          {
            "name": "Filters",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Sorts",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Page",
            "in": "query",
            "schema": {
              "maximum": 2147483647,
              "minimum": 1,
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "PageSize",
            "in": "query",
            "schema": {
              "maximum": 2147483647,
              "minimum": 1,
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OkListResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LegalContractListDtoOkListResponse"
                }
              }
            }
          },
          "404": {
            "description": "NotFound"
          }
        }
      },
      "post": {
        "tags": [
          "LegalContracts"
        ],
        "operationId": "CreateLegalConractAsync",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LegalContractCreateOptions"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/LegalContractCreateOptions"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/LegalContractCreateOptions"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "CreatedResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LegalContractDetailDtoOkListResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/legal-contracts/{id}": {
      "get": {
        "tags": [
          "LegalContracts"
        ],
        "operationId": "GetLegalContractByIdAsync",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OkListResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LegalContractDetailDtoOkListResponse"
                }
              }
            }
          },
          "404": {
            "description": "NotFound"
          }
        }
      },
      "patch": {
        "tags": [
          "LegalContracts"
        ],
        "operationId": "UpdateLegalConractAsync",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LegalContractUpdateOptions"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/LegalContractUpdateOptions"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/LegalContractUpdateOptions"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "UpdatedResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LegalContractDetailDtoOkListResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "LegalContracts"
        ],
        "operationId": "DeleteLegalContractById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "NotFound"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "LegalContractCreateOptions": {
        "required": [
          "author",
          "title"
        ],
        "type": "object",
        "properties": {
          "author": {
            "maxLength": 128,
            "minLength": 1,
            "type": "string",
            "description": "The name of the author of the contract",
            "example": "Mr. John Doe"
          },
          "title": {
            "maxLength": 256,
            "minLength": 1,
            "type": "string",
            "description": "The title of the contract",
            "example": "Sale of a car"
          },
          "description": {
            "maxLength": 512,
            "type": "string",
            "description": "A short description of the legal contract",
            "nullable": true,
            "example": "The contract is about the sale of a car, the car is a 2020 model and has a mileage of 1000km, the car is in good condition."
          }
        },
        "additionalProperties": false
      },
      "LegalContractDetailDto": {
        "required": [
          "author",
          "title"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "author": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "modifiedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "LegalContractDetailDtoOkListResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "The HTTP response status codes",
            "format": "int32",
            "nullable": true
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LegalContractDetailDto"
            },
            "nullable": true
          },
          "description": {
            "type": "string",
            "description": "A text describing the response",
            "nullable": true
          },
          "message": {
            "type": "string",
            "description": "A response message delivered with the response",
            "nullable": true
          },
          "count": {
            "type": "integer",
            "format": "int32"
          },
          "totalCount": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "LegalContractListDto": {
        "required": [
          "author",
          "title"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "author": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "modifiedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "LegalContractListDtoOkListResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "The HTTP response status codes",
            "format": "int32",
            "nullable": true
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LegalContractListDto"
            },
            "nullable": true
          },
          "description": {
            "type": "string",
            "description": "A text describing the response",
            "nullable": true
          },
          "message": {
            "type": "string",
            "description": "A response message delivered with the response",
            "nullable": true
          },
          "count": {
            "type": "integer",
            "format": "int32"
          },
          "totalCount": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "LegalContractUpdateOptions": {
        "required": [
          "author",
          "title"
        ],
        "type": "object",
        "properties": {
          "author": {
            "maxLength": 128,
            "minLength": 1,
            "type": "string",
            "description": "The name of the author of the contract",
            "example": "Mr. John Doe"
          },
          "title": {
            "maxLength": 256,
            "minLength": 1,
            "type": "string",
            "description": "The title of the contract",
            "example": "Sale of a car"
          },
          "description": {
            "maxLength": 512,
            "type": "string",
            "description": "A short description of the legal contract",
            "nullable": true,
            "example": "The contract is about the sale of a car, the car is a 2020 model and has a mileage of 1000km, the car is in good condition."
          }
        },
        "additionalProperties": false
      }
    }
  }
}